fixbug mvcc list
