+++ blockchain/localdb.go
@@ -1,8 +1,6 @@
 package blockchain
 
 import (
-	"sync"
-
 	"github.com/33cn/chain33/common"
 	"github.com/33cn/chain33/common/db"
 	"github.com/33cn/chain33/queue"
@@ -83,7 +81,7 @@ func (chain *BlockChain) localSet(msg *queue.Message) {
 
 //创建 localdb transaction
 func (chain *BlockChain) localNew(msg *queue.Message) {
-	tx := NewLocalDB(chain.blockStore.db)
+	tx := db.NewLocalDB(chain.blockStore.db)
 	id := common.StorePointer(tx)
 	msg.Reply(chain.client.NewMessage("", types.EventLocalNew, &types.Int64{Data: id}))
 }
@@ -154,161 +152,4 @@ func (chain *BlockChain) localPrefixCount(msg *queue.Message) {
 	Prefix := (msg.Data).(*types.ReqKey)
 	counts := db.NewListHelper(chain.blockStore.db).PrefixCount(Prefix.Key)
 	msg.Reply(chain.client.NewMessage("", types.EventLocalReplyValue, &types.Int64{Data: counts}))
-}
-
-// LocalDB local db for store key value in local
-type LocalDB struct {
-	txcache db.DB
-	cache   db.DB
-	maindb  db.DB
-	intx    bool
-	mu      sync.RWMutex
-}
-
-func newMemDB() db.DB {
-	memdb, err := db.NewGoMemDB("", "", 0)
-	if err != nil {
-		panic(err)
-	}
-	return memdb
-}
-
-// NewLocalDB new local db
-func NewLocalDB(maindb db.DB) db.KVDB {
-	return &LocalDB{
-		cache:   newMemDB(),
-		txcache: newMemDB(),
-		maindb:  maindb,
-	}
-}
-
-// Get get value from local db
-func (l *LocalDB) Get(key []byte) ([]byte, error) {
-	l.mu.RLock()
-	defer l.mu.RUnlock()
-	value, err := l.get(key)
-	return value, err
-}
-
-func (l *LocalDB) get(key []byte) ([]byte, error) {
-	if l.intx && l.txcache != nil {
-		if value, err := l.txcache.Get(key); err == nil {
-			return value, nil
-		}
-	}
-	if value, err := l.cache.Get(key); err == nil {
-		return value, nil
-	}
-	value, err := l.maindb.Get(key)
-	if err != nil {
-		return nil, err
-	}
-	err = l.cache.Set(key, value)
-	if err != nil {
-		panic(err)
-	}
-	return value, nil
-}
-
-// Set set key value to local db
-func (l *LocalDB) Set(key []byte, value []byte) error {
-	l.mu.Lock()
-	defer l.mu.Unlock()
-	if l.intx {
-		if l.txcache == nil {
-			l.txcache = newMemDB()
-		}
-		setdb(l.txcache, key, value)
-	} else {
-		setdb(l.cache, key, value)
-	}
-	return nil
-}
-
-// List 从数据库中查询数据列表，set 中的cache 更新不会影响这个list
-func (l *LocalDB) List(prefix, key []byte, count, direction int32) ([][]byte, error) {
-	l.mu.RLock()
-	defer l.mu.RUnlock()
-	dblist := make([]db.IteratorDB, 0)
-	if l.txcache != nil {
-		dblist = append(dblist, l.txcache)
-	}
-	if l.cache != nil {
-		dblist = append(dblist, l.cache)
-	}
-	if l.maindb != nil {
-		dblist = append(dblist, l.maindb)
-	}
-	mergedb := db.NewMergedIteratorDB(dblist)
-	it := db.NewListHelper(mergedb)
-	return it.List(prefix, key, count, direction), nil
-}
-
-// PrefixCount 从数据库中查询指定前缀的key的数量
-func (l *LocalDB) PrefixCount(prefix []byte) (count int64) {
-	l.mu.RLock()
-	defer l.mu.RUnlock()
-	dblist := make([]db.IteratorDB, 0)
-	if l.txcache != nil {
-		dblist = append(dblist, l.txcache)
-	}
-	if l.cache != nil {
-		dblist = append(dblist, l.cache)
-	}
-	if l.maindb != nil {
-		dblist = append(dblist, l.maindb)
-	}
-	mergedb := db.NewMergedIteratorDB(dblist)
-	it := db.NewListHelper(mergedb)
-	return it.PrefixCount(prefix)
-}
-
-//Begin 开启内存事务处理
-func (l *LocalDB) Begin() {
-	l.mu.Lock()
-	defer l.mu.Unlock()
-	l.intx = true
-	l.txcache = nil
-}
-
-// Rollback reset tx
-func (l *LocalDB) Rollback() {
-	l.mu.Lock()
-	defer l.mu.Unlock()
-	l.resetTx()
-}
-
-// Commit canche tx
-func (l *LocalDB) Commit() error {
-	l.mu.Lock()
-	defer l.mu.Unlock()
-	if l.txcache == nil {
-		l.resetTx()
-		return nil
-	}
-	it := l.txcache.Iterator(nil, nil, false)
-	for it.Next() {
-		l.cache.Set(it.Key(), it.Value())
-	}
-	l.resetTx()
-	return nil
-}
-
-func (l *LocalDB) resetTx() {
-	l.intx = false
-	l.txcache = nil
-}
-
-func setdb(d db.DB, key []byte, value []byte) {
-	if value == nil {
-		err := d.Delete(key)
-		if err != nil {
-			return
-		}
-	} else {
-		err := d.Set(key, value)
-		if err != nil {
-			panic(err)
-		}
-	}
 }
