+++ common/db/mvcc_iter_test.go
@@ -117,12 +117,12 @@ func TestGetAllCoinsMVCCIter(t *testing.T) {
 	}
 
 	fmt.Println("---case 1-2----")
-	var match_values [][]byte
+	var matchValues [][]byte
 	listhelper.IteratorCallback([]byte("mavl-coins-bty-"), nil, 0, 1, func(key, value []byte) bool {
-		match_values = append(match_values, value)
+		matchValues = append(matchValues, value)
 		return false
 	})
-	values = match_values
+	values = matchValues
 	assert.Equal(t, "3", string(values[0]))
 	assert.Equal(t, "2", string(values[1]))
 	assert.Equal(t, "4", string(values[2]))
@@ -150,12 +150,12 @@ func TestGetAllCoinsMVCCIter(t *testing.T) {
 	}
 	//m.PrintAll()
 	fmt.Println("---case 2-2----")
-	match_values = nil
+	matchValues = nil
 	listhelper.IteratorCallback(([]byte("mavl-coins-bty-")), []byte("mavl-coins-bty-exec-"), 0, 1, func(key, value []byte) bool {
-		match_values = append(match_values, value)
+		matchValues = append(matchValues, value)
 		return false
 	})
-	values = match_values
+	values = matchValues
 
 	for i := 0; i < len(values); i++ {
 		fmt.Println(string(values[i]))
@@ -166,12 +166,12 @@ func TestGetAllCoinsMVCCIter(t *testing.T) {
 	assert.Equal(t, "4", string(values[2]))
 
 	fmt.Println("---case 2-3----")
-	match_values = nil
+	matchValues = nil
 	listhelper.IteratorCallback(([]byte("mavl-coins-bty-16htvcBNSEA7fZhAdLJphDwQRQJaHpyHTq")), []byte("mavl-coins-bty-exec-"), 0, 1, func(key, value []byte) bool {
-		match_values = append(match_values, value)
+		matchValues = append(matchValues, value)
 		return false
 	})
-	values = match_values
+	values = matchValues
 
 	for i := 0; i < len(values); i++ {
 		fmt.Println(string(values[i]))
@@ -180,4 +180,50 @@ func TestGetAllCoinsMVCCIter(t *testing.T) {
 	//assert.Equal(t, "3", string(values[0]))
 	assert.Equal(t, "2", string(values[0]))
 	assert.Equal(t, "4", string(values[1]))
+}
+
+func TestSimpleMVCCLocalDB(t *testing.T) {
+	//use leveldb
+	//use localdb
+	db3, dir := newGoLevelDB(t)
+	defer os.RemoveAll(dir) // clean up
+	kvdb := NewLocalDB(db3)
+	m := NewSimpleMVCC(kvdb)
+
+	kvlist, err := m.AddMVCC(KeyValueList([2]string{"mavl-coins-bty-exec-16htvcBNSEA7fZhAdLJphDwQRQJaHpyHTp", "1"}, [2]string{"mavl-coins-bty-16htvcBNSEA7fZhAdLJphDwQRQJaHpyHTq", "2"}), hashN(0), nil, 0)
+	assert.Nil(t, err)
+	setKVList(kvdb, kvlist)
+
+	kvlist, err = m.AddMVCC(KeyValueList([2]string{"mavl-coins-bty-16htvcBNSEA7fZhAdLJphDwQRQJaHpyHTp", "3"}, [2]string{"mavl-coins-bty-16htvcBNSEA7fZhAdLJphDwQRQJaHpyHTs", "4"}), hashN(1), hashN(0), 1)
+	assert.Nil(t, err)
+	setKVList(kvdb, kvlist)
+
+	kvlist, err = m.AddMVCC(KeyValueList([2]string{"mavl-coins-bty-exec-16htvcBNSEA7fZhAdLJphDwQRQJaHpyHTp", "5"}, [2]string{"mavl-coins-bty-16htvcBNSEA7fZhAdLJphDwQRQJaHpyHTt", "6"}), hashN(2), hashN(1), 2)
+	assert.Nil(t, err)
+	setKVList(kvdb, kvlist)
+
+	values, err := kvdb.List([]byte("mavl-coins-bty-"), nil, 100, 1)
+	assert.Nil(t, err)
+	assert.Equal(t, "3", string(values[0]))
+	assert.Equal(t, "2", string(values[1]))
+	assert.Equal(t, "4", string(values[2]))
+	assert.Equal(t, "6", string(values[3]))
+	assert.Equal(t, "5", string(values[4]))
+
+	kvlist, err = m.DelMVCC(hashN(2), 2, true)
+	assert.Nil(t, err)
+	setKVList(kvdb, kvlist)
+	values, err = kvdb.List(nil, nil, 0, 1)
+	assert.Nil(t, err)
+	assert.Equal(t, 4, len(values))
+	assert.Equal(t, "3", string(values[0]))
+	assert.Equal(t, "2", string(values[1]))
+	assert.Equal(t, "4", string(values[2]))
+	assert.Equal(t, "1", string(values[3]))
+}
+
+func setKVList(db KVDB, kvlist []*types.KeyValue) {
+	for _, v := range kvlist {
+		db.Set(v.Key, v.Value)
+	}
 }
